Modularize the codegen scripts for easier development.
